requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.23.0,<0.24.0
python-dotenv>=1.0.0
supabase>=1.0.0
plotly>=5.24.1
